Renamed `component` folder to `jt-component`
I did this because it now contains a class with the name `JtComponent`
