+++ lib/countdown-timer/countdown-timer.js
@@ -1,6 +1,6 @@
 /* globals CustomEvent */
 
-import { JtComponent } from '../component/main.js'
+import { JtComponent } from '../jt-component/main.js'
 
 /**
  * Custom component which creates a countdown timer. The following attributes
