+++ lib/editable-list/editable-list.js
@@ -1,4 +1,4 @@
-import { JtComponent } from '../component/main.js'
+import { JtComponent } from '../jt-component/main.js'
 
 /**
  * Custom component which creates a list to which the user can add or remove
