+++ lib/jt-component/element-map.js
