+++ lib/jt-component/main.js
