+++ lib/jt-component/template.js
